libft

+++ libft/libc/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: mlaouedj <mlaouedj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 15:04:49 by mlaouedj          #+#    #+#             */
-/*   Updated: 2020/04/30 16:45:29 by mlaouedj         ###   ########.fr       */
+/*   Updated: 2020/05/06 16:33:22 by mlaouedj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_strlcpy(char *dst, char *src, size_t n)
 {
-	unsigned int i;
-	int count;
+	size_t		i;
+	int			count;
 
 	i = 0;
 	count = 0;
